+++ lionagi/core/__init__.py
@@ -1,8 +1,8 @@
-from .messages import System, Instruction
-from .instruction_set import InstructionSet
-from .conversation import Conversation
-from .branch import Branch
-from .sessions import Session
+from .messages.messages import System, Instruction
+from .instruction_set.instruction_set import InstructionSet
+from .branch.conversation import Conversation
+from .branch.branch import Branch
+from .sessions.session import Session
 
 
 __all__  = [
