+++ lionagi/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.202"
+__version__ = "0.0.203"
